Fix for scenario in which type gets wiped by annotation (#179)
* Asserting that field with default is required

because the explicitly specified required setting in the annotation
should trump whatever the required setting is based on the type.

The only case in which we can't trust the annotation is when `required =
false`, because that's the default parameter value.

branch:

* Add reproduction scenario for incorrect type

When a field is annotated, its type is ignored. Even if the annotation
doesn't explicitly specify a type i.e. when either `implementation` or
`type` are set.

* Fix incorrect type override by annotation

* Add test cases for optional defaults

branch:

* Only assert on erased type when Scala < 3

branch: